query page with mail API, Update Contact

+++ src/main/java/p/contact/manager/repository/ContactRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import p.contact.manager.entity.Contact;
 
-public interface ContactRepository extends JpaRepository<Contact, Long> {
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
 
 	@Query("select u from Contact u where u.customer.id =:userId")
 	public Page<Contact> getContactByUser(@Param("userId") Long userId, Pageable pageable);
